+++ x/gov/migrations/v3/convert.go
@@ -3,6 +3,8 @@ package v3
 import (
 	"fmt"
 
+	"github.com/cosmos/gogoproto/proto"
+
 	"cosmossdk.io/math"
 
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
@@ -49,8 +51,18 @@ func ConvertToLegacyProposal(proposal v1.Proposal) (v1beta1.Proposal, error) {
 	if err != nil {
 		return v1beta1.Proposal{}, err
 	}
-	if len(msgs) != 1 {
-		return v1beta1.Proposal{}, sdkerrors.ErrInvalidType.Wrap("can't convert a gov/v1 Proposal to gov/v1beta1 Proposal when amount of proposal messages not exactly one")
+	if len(msgs) == 0 {
+		// If there is no messages, consider proposal as a text proposal
+		content := v1beta1.NewTextProposal(proposal.Title, proposal.Summary)
+		msg, ok := content.(proto.Message)
+		if !ok {
+			return v1beta1.Proposal{}, sdkerrors.ErrInvalidType.Wrap("can't convert a gov/v1 Proposal to gov/v1beta1 Proposal: content is not a proto message")
+		}
+		legacyProposal.Content, err = codectypes.NewAnyWithValue(msg)
+		return legacyProposal, err
+	}
+	if len(msgs) > 1 {
+		return v1beta1.Proposal{}, sdkerrors.ErrInvalidType.Wrap("can't convert a gov/v1 Proposal to gov/v1beta1 Proposal when amount of proposal messages exceeds one")
 	}
 	if legacyMsg, ok := msgs[0].(*v1.MsgExecLegacyContent); ok {
 		// check that the content struct can be unmarshalled